pinetime: avoid using an edge interrupt for touch events
This saves about 0.47mA. I'm not sure why edge interrupts use so much
power, but there is a simple workaround that uses barely any power: the
pin sense mechanism with the LATCH register that doesn't involve
interrupts at all.
It's a bit of a hack, but it works for now.