all: add testing for API conformance.
This makes it much harder for new board-specific APIs to be added.

Board specific APIs are not allowed because the goal of the board
package is to provide an API that can be used for all code. If some
features are not available on a given board they should instead be
implemented using dummy values, like NoBattery in board.Power.